refactor(linter): move expansion analysis into facts
