+++ crates/shuck-linter/src/rules/common/mod.rs
@@ -1,5 +1,4 @@
 pub mod command;
-pub mod expansion;
 pub mod query;
 pub mod safe_value;
 pub mod word;
